refactor(coordination): extract helper functions and optimize database and logging status tracking in graph_admin
